refactor(utils): extract formatRelativeTime into reusable date utility
- Moved inline date formatter from status component to src/lib/utils/formatDate.ts.
- Simplifies component logic and enables reuse across the app.
