+++ src/lib/utils/formatDate.ts
@@ -1,0 +1,28 @@
+/**
+ * Formats a date string into a relative human-readable time.
+ * Examples:
+ * - "just now"
+ * - "5m ago"
+ * - "2h ago"
+ * - "3d ago"
+ * - "14 Oct" or "14 Oct 2024" (if year differs)
+ */
+export function formatRelativeTime(dateString: string): string {
+	const date = new Date(dateString);
+	const now = new Date();
+	const diffMs = now.getTime() - date.getTime();
+	const diffMins = Math.floor(diffMs / 60000);
+	const diffHours = Math.floor(diffMins / 60);
+	const diffDays = Math.floor(diffHours / 24);
+
+	if (diffMins < 1) return 'just now';
+	if (diffMins < 60) return `${diffMins}m ago`;
+	if (diffHours < 24) return `${diffHours}h ago`;
+	if (diffDays < 7) return `${diffDays}d ago`;
+
+	return date.toLocaleDateString('en-GB', {
+		day: 'numeric',
+		month: 'short',
+		year: date.getFullYear() !== now.getFullYear() ? 'numeric' : undefined
+	});
+}
